- Protected ResourceManager.claimIOPorts using ResourcePermission - Added protocol factory plugin - Added US-ASCII decoder fix by Markhale - Removed some warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@301 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ net/src/driver/org/jnode/driver/net/lance/InitializationBlock32Bit.java
@@ -33,7 +33,7 @@ public class InitializationBlock32Bit {
 		this.offset = offset;
 
 		// Populate the initial data structure
-		mem.setShort(offset + 0x00, (short) mode);
+		mem.setShort(offset + 0x00, mode);
 		mem.setByte(offset + 0x02, getEncodedRingLength(rxRing.getLength()));
 		mem.setByte(offset + 0x03, getEncodedRingLength(txRing.getLength()));
 		mem.setByte(offset + 0x04, physicalAddr.get(0));
@@ -44,8 +44,8 @@ public class InitializationBlock32Bit {
 		mem.setByte(offset + 0x09, physicalAddr.get(5));
 		mem.setInt(offset + 0x0C, (int) (logicalAddr & 0xFFFFFFFF));
 		mem.setInt(offset + 0x10, (int) ((logicalAddr >> 32) & 0xFFFFFFFF));
-		mem.setInt(offset + 0x14, (int) rxRing.getAddressAs32());
-		mem.setInt(offset + 0x18, (int) txRing.getAddressAs32());
+		mem.setInt(offset + 0x14, rxRing.getAddressAs32());
+		mem.setInt(offset + 0x18, txRing.getAddressAs32());
 	}
 
 	private byte getEncodedRingLength(int ringLength) {
